Automatically backed up by Learn

+++ app/controllers/static_controller.rb
@@ -10,7 +10,7 @@ class StaticController < ApplicationController
   end
 
   def logout
-    session.delete(:user_id)
+    session.delete(:user_id) if session[:user_id]
     redirect_to root_url
   end
 end
